+++ stroom-core-client/src/main/java/stroom/index/client/presenter/IndexVolumeStatusListPresenter.java
@@ -202,7 +202,6 @@ public class IndexVolumeStatusListPresenter extends MyPresenterWidget<PagerView>
                         .enabledWhen(this::isEnabled)
                         .build(),
                 DataGridUtil.headingBuilder("Usage Date")
-                        .centerAligned()
                         .withToolTip("The date/time this volume was last written to.")
                         .build(),
                 ColumnSizeConstants.DATE_AND_DURATION_COL);